skinalyze-ai/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

# --- AI Core ---
langchain==0.1.16
langchain-community==0.0.33
langchain-google-genai==0.0.11
langchain-chroma==0.1.0
langchain-core<0.2.0
google-generativeai==0.4.1
chromadb==0.4.22

# --- Vector Store Dependencies (FIX LỖI 1) ---
sentence-transformers>=2.2.2
scikit-learn
huggingface-hub

# --- Torch CPU ---
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cpu

# --- Image Processing ---
Pillow
numpy
opencv-python-headless
mediapipe
# SAM2
git+https://github.com/facebookresearch/segment-anything-2.git

# --- Utilities ---
pandas